42nd.  Finished Website Security Assignment.

+++ marathon/app/Controllers/Admin.php
@@ -2,6 +2,7 @@
 
 namespace App\Controllers;
 
+use App\Models\Member;
 use App\Models\Race;
 
 class Admin extends BaseController
@@ -14,6 +15,7 @@ class Admin extends BaseController
     //Navigation
     public function manage_marathon(): string //added from week 13 in class work
     {
+
         $Race=new Race();
         $data= array('manage_marathon' => 'true');
         $data['races'] = $Race->get_races();
@@ -46,24 +48,67 @@ class Admin extends BaseController
     public function delete_race($id)
     {
         $Race=new Race();
-        $Race->delete_races($id);
+        $this->session = service('session');
+        $this->session->start();
+        $memberKey = $this->session->get("memberKey");
+        $Member = new Member();
+        if($Member->has_access($id, $memberKey))
+        {
+
+            $Race->delete_races($id);
+
+
+        }else{
+            echo "Error Access Denied";
+            exit();
+        }
+
+
         header("Refresh:0; url=/marathon/public/marathon");
         exit();
     }
     //Update Race: loading view with the data that needs updating
     public function update_race($id)
     {
-        $Race=new Race();
         $data= array('manage_marathon' => 'true');
-        $data['race'] = $Race->get_race($id);
+        $this->session = service('session');
+        $this->session->start();
+        $memberKey = $this->session->get("memberKey");
+        $Member = new Member();
+        if($Member->has_access($id, $memberKey))
+        {
+            $Race=new Race();
 
+            $data['race'] = $Race->get_race($id);
+
+
+        }else{
+            echo "Error Access Denied";
+            exit();
+        }
         return view("update_page",$data);
+
 
     }
     public function edit_race()
     {
         $Race=new Race();
-        $Race->update_races($this->request->getPost('Name'),$this->request->getPost('Location'),$this->request->getPost('Description'),$this->request->getPost('Date'),$this->request->getPost('txtID'));
+
+        $this->session = service('session');
+        $this->session->start();
+        $memberKey = $this->session->get("memberKey");
+        $Member = new Member();
+        if($Member->has_access('txtID', $memberKey))
+        {
+            $Race->update_races($this->request->getPost('Name'),$this->request->getPost('Location'),$this->request->getPost('Description'),$this->request->getPost('Date'),$this->request->getPost('txtID'));
+
+
+        }else{
+            echo "Error Access Denied";
+            exit();
+        }
+
+
         header("Location: /marathon/public/marathon");
         exit();
